Adjust unit tests for new features. Disabled ones for stuff that hasn't been implemented yet so CI passes
